requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
